fix(indent): fix indent cache tracking
There are places that track indents as strings and places that track
them as numeric values. In the case of ignored tokens it would result in
numeric values being added with string values `00010`. This makes sure
the values in the indent cache are always numeric based on the indent
size

fixes: gh-9